+++ objc2/CHANGELOG.md
@@ -50,6 +50,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * **BREAKING**: Changed `MessageReceiver::send_message` to panic instead of
   returning an error.
 
+### Fixed
+* **BREAKING**: Disallow throwing `nil` exceptions in `exception::throw`.
+
 ### Removed
 * **BREAKING**: Removed the `Sel::from_ptr` and `Sel::as_ptr` methods.
 * **BREAKING**: Removed `MessageError`.